requirements.txt
discord.py
paramiko
aiosqlite
python-dotenv